Trigger autosave after rupee decrement is complete instead of omitting it entirely

+++ soh/include/z64save.h
@@ -197,7 +197,7 @@ typedef struct {
     u16 adultTradeItems;
     u8 pendingIceTrapCount;
     u8 mqDungeonCount;
-    u8 pendingSale;
+    u16 pendingSale;
     SohStats sohStats;
 } SaveContext; // size = 0x1428
